Add file check / model audit tool.
Searches through all of the model files in a parent folder and
checks that the files referenced by the model files exist in
the location they should be.

If the file is not where it should be (incorrect path) it checks
to see if it can be found somewhere else.

Results are output as a summary of the search, a list of missing
files and a list of files found elsewhere.

User can review all of the files that reference the missing /
found elsewhere files and view the contents of the file that
references it.