Add multiple tsf summary to run variables check.
User can select a root folder and the tool will walk all of the
subfolders to find TUFLOW .tsf files.

The contents of the .tsf files are read in and added to a summary
table with key attributes from the simulation.

Right-click context menu allows the user to view more detail
by loading the associated .tlf file (referenced in the .tsf) and
load that in the variables / diagnostics tabs. The .tlf is
assumed to be in the same folder as the .tsf file (I think this
should always be the case?).