[Bugfix] Variable check fail on steady state fmp runs.
Checking for timestep value for all FMP .ief files and expecting
it to always be there. This is not the case for steady state runs.

Add an additional check to see if the "RunType" value is
"unsteady" if it is it grabs the timestep. If it isn't it will
say "Not Found" and add the RunType to the description.

Add additional error catching for the multiple .ief loader to
fail on the individual .ief file and continue. After loading
it notifies the user about which files failed.