[Bugfix] TUFLOW nwk attributes can be different from empty file.
TUFLOW will accept a 1d_nwk line file that contains only a single
column that it assumes to be the fmp node name.

Now checks to see if there's an 'ID' column name and uses that if
found, otherwise defaults to taking the first column.
Checks to see if there is a "Len_or_ANA" column and, if not found,
sets the value to 0 (default) and ignores it. Otherwise it uses
the value like usual.