Improve FMP section check layout and graphing behaviour.
Changes the layout of the section check tool dialog to add a
layout alongside the conveyance and banktop check tables - which
are now placed in a tab widget. When the user clicks on a row in
one of the tables the graph is updated and the map zooms to a
node (rather than opening a new dialog window each time).

Help file updated for new functionality.