[Bugfix] Failed reload in QGIS when using "widgets" module.
Bug when loading from zip file caused by the bootstrapper
failing on a KeyError while unloading modules.

Not sure why, but there seems to (probably) be some kind of
name clash issue when QGIS uses the bootstrapper to unload
modules from the plugin. I guess it's finding the wrong widgets
module or getting in a mess somehow because of a name clash?

Renames the widgets module to mywidgets which seems to have
fixed the problem.