Add FMP stability check tool.
Setup the stability tool to work with the FM API.

Loads the ZZN results and the DAT file in. Works the same as before,
i.e. calculates the 1st and 2nd derivate, does a bit of hacky rule
checking and identifies unstable time series.

Adds a progress bar to the dialog to let you user know what's going
on when loading a large file. The instability checks can take a
while for bigger models.

TODO:
Should add something to save the data to json for quick re-load at
some point.