Better UI updates for width check via statusbar.
Calculating the snapping of the CN lines and 1D nodes can take
quite a long time for a large model. UI can freeze during this if
no events are processed.

Adds a status update for the node being processed during the
snapping operation to avoid UI freeze.

TODO: Should implement a progress bar for this stuff really.