Adds HX line calculation support for chainage compare tool.
Previous approach could not be used if a 1d_nwk line layer was not
available. 1d_nwk lines are only required in FM models if WLLs are
used, so aren't always available.

New approach uses the HX/CN lines to calculate an approximate
distance between the nodes in 2D.
- Finds CNs snapped to 1D nodes.
- Finds the end points (where HX should be).
- Calculates HX length between CN snapped points.
- Averages the length of both HX lines (if found) to get length.
- Compares the average to the FM distance to next.

Still Beta:
- Some strange results coming up still.
- Algorithm is not great. Very slow and needs optimising.