Big improvement to tlf read times and fixed table layouts.
Add check to the tlf loaded to skip line if the read is within
the simulation output section. Nothing is needed from this part
and it can take up a lot of the file. Sets a flag to skip lines
that will be True until the summary section is found.

Fit all table data to contents to make the layouts a bit nicer.

Add a column to the summary table to show whether any warnings
were found.