[Bugfix] entry_tcf can be only a name in the .tlf file.
In some cases (older models?) the entry_tcf can have only a
filename with no path.

Add support to find the 'home' folder in the tlf and check
whether the entry_tcf is a filename or an absolute path. If a
filename is found, the home folder path will be prepended to
provide a full path.

Some changes to the configuration of BC Event Source / Text
made to include it in the variables for easier lookup. Will be
useful for reading boundary condition files.